requirements.txt
discord.py>=2.3.0
asyncpg>=0.28.0
httpx>=0.24.0
python-dotenv>=1.0.0
psutil>=5.9.0

